Put back OpenAIProtocol
