+++ Tests/OpenAITests/OpenAITestsCombine.swift
@@ -10,12 +10,10 @@
 import XCTest
 @testable import OpenAI
 
-@available(iOS 13.0, *)
-@available(watchOS 6.0, *)
-@available(tvOS 13.0, *)
+@available(iOS 13.0, tvOS 13.0, macOS 10.15, watchOS 6.0, *)
 final class OpenAITestsCombine: XCTestCase {
     
-    var openAI: OpenAI!
+    var openAI: OpenAIProtocol!
     var urlSession: URLSessionMock!
     
     override func setUp() {
